requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==44.0.3
dj-rest-auth==7.0.1
Django==5.2
django-allauth==65.7.0
djangorestframework==3.16.0
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
packaging==25.0
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
psycopg2-binary==2.9.10
pycparser==2.22
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
uritemplate==4.1.1
urllib3==2.4.0
